refactor: move version printing into internal/version
