+++ internal/version/version.go
@@ -1,4 +1,4 @@
-package main
+package version
 
 import (
 	"fmt"
